api/requirements.txt
fastapi
uvicorn[standard]
websockets
redis[hiredis]
pydantic
pandas
langchain
langchain-openai
langchain-core
python-dotenv 
playwright
requests 